Fix problem with getting current volume on the ubuntu platform
The output from "pacmd list-sinks" doesn't seem to be the same as
for the platform this script was developed for. I changed it so it
works for ubuntu now and added an if statement to check which
platform this is running at. Just update the if statement to your
platform. Currently I have the original code under "Something else"
as platform name.